This change adds parameterized modules to KRL.
The many files changed are primarily because of a need to refactor how keys work so that they can be scoped inside modules.  This is the area where there is the highest likelihood of problems since so much was changed.

The parser changes are acretive, so optimization level was not changed.

Changes for parameterized modules themselves are primarily limited to Modules.pm for execution, and Rules.pm (eval_meta, eval_global) for set up of the module environment.  Module environments are stored in the regular environment.